Move extraParams to it's own directory
