+++ api/src/__tests__/support/types/extraParams/guildPlaylist.params.ts
@@ -1,0 +1,13 @@
+import { GuildResolvable } from 'discord.js';
+import { GuildPlaylist } from '../../../../entities/index.js';
+
+export interface GuildPlaylistMocks {
+    hasPermission: Function,
+    isGuildMember: Function
+};
+
+export interface GuildPlaylistExtraParams {
+    mocks: GuildPlaylistMocks,
+    guildId: GuildResolvable,
+    createdPlaylist?: GuildPlaylist
+};
