+++ api/src/__tests__/support/types/extraParams/guildPlaylistRole.params.ts
@@ -1,0 +1,13 @@
+import { GuildResolvable } from 'discord.js';
+import { GuildPlaylistRole } from '../../../../entities/index.js';
+
+export interface GuildPlaylistRoleMocks {
+    hasPermission: Function,
+    isGuildMember: Function
+};
+
+export interface GuildPlaylistRoleExtraParams {
+    mocks: GuildPlaylistRoleMocks,
+    guildId: GuildResolvable,
+    guildPlaylistRole?: GuildPlaylistRole
+};
