+++ api/src/__tests__/support/types/extraParams/guildSong.params.ts
@@ -1,0 +1,15 @@
+import { GuildResolvable } from 'discord.js';
+import { GuildSong } from '../../../../entities/index.js';
+
+export interface GuildSongMocks {
+    hasPermission: Function,
+    isGuildMember: Function,
+    hasGuildPlaylistRole: Function,
+    handleSearch: Function
+};
+
+export interface GuildSongExtraParams {
+    mocks: GuildSongMocks,
+    guildId: GuildResolvable,
+    createdSong?: GuildSong
+};
