+++ api/src/__tests__/support/types/extraParams/index.ts
@@ -1,0 +1,6 @@
+export * from './guildPlaylist.params.js';
+export * from './guildPlaylistRole.params.js';
+export * from './guildSong.params.js';
+
+export * from './userPlaylist.params.js';
+export * from './userSong.params.js';
