+++ api/src/__tests__/support/types/extraParams/userPlaylist.params.ts
@@ -1,0 +1,5 @@
+import { UserPlaylist } from '../../../../entities/index.js';
+
+export interface UserPlaylistExtraParams {
+    createdPlaylist?: UserPlaylist
+};
