+++ api/src/__tests__/support/types/extraParams/userSong.params.ts
@@ -1,0 +1,10 @@
+import { UserSong } from '../../../../entities/index.js';
+ 
+export interface UserSongMocks {
+    handleSearch: Function
+};
+
+export interface UserSongExtraParams {
+    mocks: UserSongMocks,
+    createdSong?: UserSong
+};
